Move some architecture bits into arch/
ARM defines the exception vector and startup for Cortex-M series, so we
can keep the base vectors as constant across all Cortex-M platforms, and
put the SoC's specific interrupts into the bsp instead.  That part will
come later, as a separate vector table to be concatenated in the linker
script.